+++ tests/test_sources.py
@@ -1,32 +1,39 @@
 import inspect
 
+import hypothesis
 import numpy as np
+from hypothesis import given, strategies as st
 
 import pyechelle.sources as sources
 
 
-# @given(
-#     st.floats(min_value=0.0, exclude_min=True, allow_nan=False, allow_infinity=False, ),
-#     st.floats(
-#         min_value=0.0,
-#         exclude_min=True,
-#         max_value=1.0,
-#         allow_nan=False,
-#         allow_infinity=False,
-#     ),
-#     st.integers(),
-# )
-def test_sources():
-    wl = np.linspace(300, 305, 1000, dtype=np.float)
-    available_sources = [m[0] for m in inspect.getmembers(sources, inspect.isclass) if
-                         issubclass(m[1], sources.Source) and m[0] != "Source"]
-    assert len(available_sources) > 0
-    for s in available_sources:
-        source = getattr(sources, s)()
-        sd = source.get_spectral_density(wl)
-        print(sd)
+@given(
+    st.floats(min_value=0.3, max_value=1.5, allow_nan=False, allow_infinity=False),
+    st.floats(min_value=0.0001, max_value=0.005, allow_nan=False, allow_infinity=False),
+    st.sampled_from([m[0] for m in inspect.getmembers(sources, inspect.isclass) if
+                     issubclass(m[1], sources.Source) and m[0] != "Source"])
+)
+@hypothesis.settings(deadline=None)
+def test_sources(wl, bw, source_name):
+    wavelength = np.linspace(wl, wl + bw, 1000, dtype=np.float)
+    print(f"Test {source_name}...")
+    source = getattr(sources, source_name)()
+    sd = source.get_spectral_density(wavelength)
+    assert isinstance(sd, tuple) or isinstance(sd, np.ndarray)
 
-    # s = sources.Constant()
-    # e = eff.ConstantEfficiency("test", eff=efficiency)
-    # assert e.get_efficiency(wl) == efficiency
-    # assert e.get_efficiency_per_order(wl, order)
+
+@given(
+    st.floats(min_value=0.1, max_value=50000., allow_nan=False, allow_infinity=False)
+)
+def test_rv_shift(rv):
+    wl = np.linspace(0.5, 0.5005, 1000)
+    c = 299792458.
+
+    et = sources.Etalon()
+    sd0 = et.get_spectral_density_rv(wl, 0.)
+    sd1 = et.get_spectral_density_rv(wl, rv)
+    print(rv)
+    assert sd0[0][0] > sd1[0][0]
+    # TODO: fix this test. Right now the problem is that some edge lines are in sd0 which are not in sd1
+    # if len(sd0[0]) == len(sd1[0]):
+    #    assert np.allclose(sd0[0], sd1[0] * ((c+rv)/c))